refactor(backend): Integer ids + Project cleanup (slug/doi/issued/version removed, NomCode/License enums) [chunk 1/3]

+++ backend/src/main/java/org/catalogueoflife/editor/name/NameUsageService.java
@@ -34,21 +34,21 @@ public class NameUsageService {
   }
 
   public List<NameUsageResponse> list(long userId, long projectId, int limit, int offset) {
-    projects.requireRole(userId, projectId);
+    projects.requireRole((int) userId, (int) projectId);
     return usages.findByProject(projectId, Pagination.clampLimit(limit), Pagination.clampOffset(offset)).stream()
         .map(this::toListResponse)
         .toList();
   }
 
   public List<NameUsageResponse> search(long userId, long projectId, String q, int limit, int offset) {
-    projects.requireRole(userId, projectId);
+    projects.requireRole((int) userId, (int) projectId);
     return usages.search(projectId, q, Pagination.clampLimit(limit), Pagination.clampOffset(offset)).stream()
         .map(this::toListResponse)
         .toList();
   }
 
   public NameUsageResponse get(long userId, long projectId, long id) {
-    projects.requireRole(userId, projectId);
+    projects.requireRole((int) userId, (int) projectId);
     Project project = requireProject(projectId);
     return toResponse(requireInProject(projectId, id), project);
   }
@@ -187,7 +187,7 @@ public class NameUsageService {
   }
 
   private Project requireProject(long projectId) {
-    Project p = projectMapper.findById(projectId);
+    Project p = projectMapper.findById((int) projectId);
     if (p == null) {
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "project not found");
     }
@@ -195,7 +195,7 @@ public class NameUsageService {
   }
 
   private void requireEditor(long userId, long projectId) {
-    String role = projects.requireRole(userId, projectId);
+    String role = projects.requireRole((int) userId, (int) projectId);
     if (!role.equals(Role.OWNER.dbValue()) && !role.equals(Role.EDITOR.dbValue())) {
       throw new ResponseStatusException(HttpStatus.FORBIDDEN, "owner or editor required");
     }
